# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a RESTful API for an E-commerce platform built with Go.",
        "title": "GO Commerce API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/api/orders": {
            "get": {
                "description": "Retrieve all orders placed by the user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get Orders",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer JWT_Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.OrderResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/orders/checkout": {
            "post": {
                "description": "Place a new order with the items in the user's cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Place Order",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer + JWT_Token\u003e",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Order Input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Order"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OrderResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid Input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/auth/login": {
            "post": {
                "description": "Login to your user account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Invalid Input",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Wrong Email or password",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Register a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "parameters": [
                    {
                        "description": "Registration credential",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Invalid Input",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.LoginRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "models.Order": {
            "description": "An order contains details like user ID, total price, status, and the time it was placed.",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Order ID",
                    "type": "integer",
                    "example": 1
                },
                "placed_at": {
                    "description": "Timestamp when the order was placed",
                    "type": "string",
                    "example": "2023-08-01T12:00:00Z"
                },
                "status": {
                    "description": "Status of the order",
                    "type": "string",
                    "example": "Pending"
                },
                "total_price": {
                    "description": "Total Price of the order",
                    "type": "number",
                    "example": 299.99
                },
                "user_id": {
                    "description": "User ID",
                    "type": "integer",
                    "example": 101
                }
            }
        },
        "models.OrderResponse": {
            "description": "Details of an order in the response.",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Order ID",
                    "type": "integer",
                    "example": 1
                },
                "items": {
                    "description": "List of ordered items",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OrderedItem"
                    }
                },
                "placed_at": {
                    "description": "Timestamp when the order was placed",
                    "type": "string",
                    "example": "2023-08-01T12:00:00Z"
                },
                "status": {
                    "description": "Status of the order",
                    "type": "string",
                    "example": "Pending"
                },
                "total_price": {
                    "description": "Total Price of the order",
                    "type": "number",
                    "example": 299.99
                },
                "user_id": {
                    "description": "User ID",
                    "type": "integer",
                    "example": 101
                }
            }
        },
        "models.OrderedItem": {
            "description": "An ordered item contains details about the product, quantity, price, and the order it belongs to.",
            "type": "object",
            "properties": {
                "id": {
                    "description": "Ordered Item ID",
                    "type": "integer",
                    "example": 1
                },
                "order_id": {
                    "description": "Order ID",
                    "type": "integer",
                    "example": 1
                },
                "price": {
                    "description": "Price of the product at the time of the order",
                    "type": "number",
                    "example": 149.99
                },
                "product_id": {
                    "description": "Product ID",
                    "type": "integer",
                    "example": 101
                },
                "quantity": {
                    "description": "Quantity of the product",
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "models.RegisterRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Optional / new fields",
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "last_login": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "description": "user or admin",
                    "type": "string"
                },
                "status": {
                    "description": "active, inactive, suspended",
                    "type": "string"
                }
            }
        }
    }
}